requirements.txt
google-generativeai
flask
flask-cors
requests
rend
